Development
Built up the CrimeFragment view. Added a checkbox and button with the date on it.  Also debugged the error about the date button being set on a null object.  It was because I forgot to add the mDate = new Date(); variable in the Crime class file.

Stopped on page 158.
Changes to be committed:
    modified:   .idea/misc.xml
    modified:   app/src/main/java/com/example/android/criminalintent/Crime.java
    modified:   app/src/main/java/com/example/android/criminalintent/CrimeFragment.java
    modified:   app/src/main/res/layout/fragment_crime.xml
    modified:   app/src/main/res/values/strings.xml