Fix
Fixed an UI issue with the page swiper.  The bug was when the page swiper was opened it always started at the first crime, no matter what crime was clicked.  Added a check at the end of onCreate that sets the index to match the index of the crime selected by matching the value to the intent extra.

Changes to be committed:
    modified:   .idea/misc.xml
    modified:   app/src/main/java/com/example/android/criminalintent/CrimePagerActivity.java