Development
Passing data between the two fragments.  Added code in DatePickerFragment to pass data to that fragment.  Changed code to call it in CrimeFragment.
Added code into DatePcikerFragment to extract the date and initiale the datepicker. Set a target fragment in CrimeFragment and sending data to the target fragment.

Changes to be committed:
    modified:   app/build.gradle
    modified:   app/src/main/java/com/example/android/criminalintent/CrimeFragment.java
    modified:   app/src/main/java/com/example/android/criminalintent/DatePickerFragment.java